+++ CMakeLists.txt
@@ -13,7 +13,5 @@ endif()
 ament_auto_package(
   INSTALL_TO_SHARE
   launch
-  maps
   params
-  rviz
 )
